feature/jenkinsfile added parallel tests
