+++ Jenkinsfile
@@ -12,19 +12,43 @@ pipeline{
             }
         }
         stage('Test') {
-            agent {
-                docker {
-                    image 'qnib/pytest'
+
+            parallel {
+                stage('on centos') {
+                    agent {
+                        docker {
+                            image 'qnib/pytest'
+                        }
+                    }
+                    steps {
+                        sh 'py.test --verbose --junit-xml test-reports/results.xml sources/test_calc.py || true'
+                    }
+                    post {
+                        always {
+                            junit 'test-reports/results.xml'
+                        }
+                    }
                 }
-            }
-            steps {
-                sh 'py.test --verbose --junit-xml test-reports/results.xml sources/test_calc.py || true'
-            }
-            post {
-                always {
-                    junit 'test-reports/results.xml'
+                stage('on debian') {
+                    agent {
+                        docker {
+                            image 'qnib/pytest'
+                        }
+                    }
+                    steps {
+                        sh 'py.test --verbose --junit-xml test-reports/results.xml sources/test_calc2.py || true'
+                    }
+                    post {
+                        always {
+                            junit 'test-reports/results.xml'
+                        }
+                    }
                 }
-            }
+//            steps {
+//                sh 'py.test --verbose --junit-xml test-reports/results.xml sources/test_calc.py || true'
+//            }
+
+        }
         }
         stage('Deliver') {
             agent {
