+++ sources/test_calc.py
@@ -8,27 +8,27 @@ class TestCalc(unittest.TestCase):
     # def test_fail(self):
     #     self.assertFalse(True, msg="This is failed test")
 
-    def test_add_integers(self):
-        """
-        Test that the addition of two integers returns the correct total
-        """
-        result = calc.add2(1, 2)
-        self.assertEqual(result, 3)
-
-    def test_add_floats(self):
-        """
-        Test that the addition of two floats returns the correct result
-        """
-        result = calc.add2('10.5', 2)
-        self.assertEqual(result, 12.5)
-
-    def test_add_strings(self):
-        """
-        Test the addition of two strings returns the two strings as one
-        concatenated string
-        """
-        result = calc.add2('abc', 'def')
-        self.assertEqual(result, 'abcdef')
+    # def test_add_integers(self):
+    #     """
+    #     Test that the addition of two integers returns the correct total
+    #     """
+    #     result = calc.add2(1, 2)
+    #     self.assertEqual(result, 3)
+    #
+    # def test_add_floats(self):
+    #     """
+    #     Test that the addition of two floats returns the correct result
+    #     """
+    #     result = calc.add2('10.5', 2)
+    #     self.assertEqual(result, 12.5)
+    #
+    # def test_add_strings(self):
+    #     """
+    #     Test the addition of two strings returns the two strings as one
+    #     concatenated string
+    #     """
+    #     result = calc.add2('abc', 'def')
+    #     self.assertEqual(result, 'abcdef')
 
     def test_add_string_and_integer(self):
         """
